newversion v7(final)/EdgeSolution/modules/modbus/requirements.txt
azure-iot-device~=2.7.0
pyModbusTCP~=0.1.10
schedule~=1.1.0
